Rename "pre 06" packages to "legacy"
